backend/requirements.txt
# requirements.txt
# Using your existing dependencies plus new features

# Core FastAPI and web server
fastapi[standard]
# uvicorn is included with fastapi[standard]
# pydantic is included with fastapi

# Machine Learning and NLP
sentence_transformers
torch
torchvision
torchaudio
transformers==4.46.3
peft
accelerate
bitsandbytes
bert_score

# Training and evaluation
trl
rouge_score

# Monitoring and visualization
wandb
plotly
plotly_express

# Lao language processing
laonlp

# HTTP client
httpx
urllib3==1.26.15

# System monitoring and utilities
psutil

# Security and Authentication
bcrypt>=4.0.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Email support
aiosmtplib>=2.0.0

# Scheduling and background tasks
apscheduler>=3.10.0

# Data validation and serialization
pydantic-settings>=2.0.0

# Testing (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0

# Development tools (optional)
black>=23.0.0
flake8>=6.0.0

# Note: sqlite3 is included with Python standard library
